chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator>=2.2.0
requests==2.32.5
flask-limiter>=3.14.0
pandas==2.1.4
